feat: Implemented Header component
